entrega 2

+++ src/main/kotlin/br/pucpr/authserver/users/controller/UserController.kt
@@ -1,13 +1,21 @@
 package br.pucpr.authserver.users.controller
 
+import br.pucpr.authserver.security.JWT
+import br.pucpr.authserver.security.UserToken
 import br.pucpr.authserver.users.SortDir
-import br.pucpr.authserver.users.User
+import br.pucpr.authserver.users.UserRepository
 import br.pucpr.authserver.users.UserService
 import br.pucpr.authserver.users.controller.requests.CreateUserRequest
+import br.pucpr.authserver.users.controller.requests.LoginRequest
+import br.pucpr.authserver.users.controller.responses.LoginResponse
 import br.pucpr.authserver.users.controller.responses.UserResponse
+import io.swagger.v3.oas.annotations.security.SecurityRequirement
 import jakarta.validation.Valid
+import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.security.core.Authentication
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -42,11 +50,24 @@ class UserController(
             ?: ResponseEntity.notFound().build()
 
     @DeleteMapping("/{id}")
-    fun delete(@PathVariable id: Long): ResponseEntity<Void> =
-        userService.delete(id)
-            .let { ResponseEntity.ok().build() }
+    @PreAuthorize("hasRole('ADMIN') or #id == principal.id")
+    @SecurityRequirement(name = "AuthServer")
+    fun delete(@PathVariable id: Long, authentication: Authentication): ResponseEntity<Void> {
+        val userToken = authentication.principal as? UserToken
+            ?: return ResponseEntity.status(HttpStatus.FORBIDDEN).build()
+
+        if (userToken.id == id || userToken.isAdmin) {
+            userService.delete(id)
+            return ResponseEntity.ok().build()
+        }
+
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build()
+    }
+
 
     @PutMapping("/{id}/roles/{role}")
+    @PreAuthorize("hasRole('USER')")
+    @SecurityRequirement(name = "AuthServer")
     fun grant(@PathVariable id: Long, @PathVariable role: String): ResponseEntity<Void> =
         if (userService.addRole(id, role)) ResponseEntity.ok().build()
         else ResponseEntity.noContent().build()
@@ -59,4 +80,9 @@ class UserController(
     @GetMapping("/{id}/courses")
     fun getCourses(@PathVariable id: Long) = userService.getCourses(id)
 
+    @PostMapping("/login")
+    fun login(@Valid @RequestBody user: LoginRequest) =
+        userService.login(user.email!!,user.password!!)
+            ?.let { ResponseEntity.ok().body(it) }
+            ?: ResponseEntity.notFound().build()
 }
